Add files, ensure files are connected
