--- src/App.js
@@ -1,29 +1,0 @@
-import logo from './logo.svg';
-import './App.css';
-
-function App() {
-  fetch("https://www.cheapshark.com/api/1.0/deals?storeID=1&upperPrice=2").then(response => {
-    return response.json()
-  })
-  .then(data => console.log(data))
-  return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
-    </div>
-  );
-}
-
-export default App;
