+++ src/components/Nav.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Nav() {
+  return (
+    <nav>
+      Oh hey Nav bar
+    </nav>
+  )
+}
+
+export default Nav;
