+++ src/components/StoreDealBox.js
@@ -1,0 +1,12 @@
+import React from "react";
+import "../styles/StoreDealBox.css";
+
+function StoreDealBox() {
+  return (
+    <div className="store-deal-box">
+      Hello Deal Box
+    </div>
+  )
+}
+
+export default StoreDealBox;
