+++ src/styles/App.css
@@ -1,0 +1,18 @@
+* {
+  margin: 0;
+  padding: 0;
+  color: white;
+}
+
+body {
+  background-color: black;
+  overflow: hidden;
+  width: 100vw;
+  height: 100vh;
+}
+
+.App {
+  width: 100%;
+  height: auto;
+}
+
